requirements.txt
pymupdf==1.24.10
faiss-cpu==1.8.0.post1
sentence-transformers==2.7.0
torch>=2.2.0,<3.0.0
langchain-core==0.3.13
pyyaml==6.0.2
numpy>=1.26
tqdm>=4.66
rich>=13.7
ollama>=0.3.0
